Test Scene
Made a duplicate of my initial test scene. Ball physics and a few bricks have been created for testing purposes. As of right now I'm not sure how to create a torus-like 2D collider to contain the ball, nor how to curve the bricks so that they can align themselves around the black hole.